services/cutline-service/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.0
pydantic-settings==2.7.0
opencv-python-headless==4.10.0.84
numpy==2.2.0
Pillow==11.1.0
rembg[cpu]==2.0.59
onnxruntime==1.20.1
